Add identity map transformation
[#983]
