+++ contrib/minitest-ruby_event_store/Gemfile
@@ -5,3 +5,4 @@ eval_gemfile '../../support/bundler/Gemfile.shared'
 
 gem 'ruby_event_store', path: '../../ruby_event_store'
 gem 'aggregate_root',   path: '../../aggregate_root'
+gem 'ruby_event_store-transformations', path: '../ruby_event_store-transformations'
