+++ contrib/ruby_event_store-transformations/Gemfile
@@ -4,4 +4,3 @@ gemspec
 eval_gemfile '../../support/bundler/Gemfile.shared'
 
 gem 'ruby_event_store', path: '../../ruby_event_store'
-gem 'activesupport', '~> 5.0'
