+++ contrib/ruby_event_store-transformations/lib/ruby_event_store/transformations.rb
@@ -2,6 +2,7 @@
 
 require_relative "transformations/version"
 require_relative "transformations/with_indifferent_access"
+require_relative "transformations/identity_map"
 
 module RubyEventStore
   module Transformations
