+++ contrib/ruby_event_store-transformations/spec/spec_helper.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
-  require '../../support/helpers/rspec_defaults'
-  require 'ruby_event_store'
-  require 'ruby_event_store/transformations'
+require 'ruby_event_store'
+require 'ruby_event_store/transformations'
+require_relative '../../../support/helpers/time_enrichment'
+require_relative '../../../support/helpers/rspec_defaults'
