+++ ruby_event_store-rspec/Gemfile
@@ -5,5 +5,5 @@ eval_gemfile '../support/bundler/Gemfile.shared'
 
 gem 'ruby_event_store', path: '../ruby_event_store'
 gem 'aggregate_root',   path: '../aggregate_root'
-
+gem 'ruby_event_store-transformations', path: '../contrib/ruby_event_store-transformations'
 gem 'diff-lcs', '= 1.3.0'
